Updated "publish docker image" action
- It was added the "workflow_call" event to make it a reusable action;
- It was added a logic to build only on the latest release (always).

The aforementioned logic is to make sure that the latest release is
selected when the action is called as a reusable workflow. The downside
is that it's not possible to select an old tag and (re-)build it. To do
it, it could be possible to create another action, with
workflow_dispatch event, which does not include the logic to build only
on the latest release.